Bugfix: detail graph
